Выделяет утилитарные функции в модуль
